% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound.R
\name{download_compound_structure}
\alias{download_compound_structure}
\title{Download compound structure file}
\usage{
download_compound_structure(client, id_type, id_value, format = "molfile")
}
\arguments{
\item{client}{mwREST client}

\item{id_type}{Identifier type ("regno", "pubchem_cid", etc.)}

\item{id_value}{Identifier value}

\item{format}{File format ("molfile" or "sdf")}
}
\value{
File content
}
\description{
Download compound structure file
}
