% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refmet.R
\name{get_all_refmet_names}
\alias{get_all_refmet_names}
\title{Get all RefMet names}
\usage{
get_all_refmet_names(client)
}
\arguments{
\item{client}{mwREST client}
}
\value{
All RefMet names
}
\description{
Get all RefMet names
}
