% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound.R
\name{get_compound_by_pubchem_cid}
\alias{get_compound_by_pubchem_cid}
\title{Get compound information by PubChem CID}
\usage{
get_compound_by_pubchem_cid(client, cid, fields = "all", format = "json")
}
\arguments{
\item{client}{An mw_rest_client object}

\item{cid}{Character or numeric. PubChem CID}

\item{fields}{Character. Fields to return (default: "all")}

\item{format}{Character. Output format ("json" or "txt")}
}
\value{
Tibble with compound information (JSON) or character string (txt)
}
\description{
Retrieve detailed compound information using a PubChem Compound ID.
}
\examples{
\dontrun{
client <- mw_rest_client()
compound_info <- get_compound_by_pubchem_cid(client, "5793")
}

}
