% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refmet.R
\name{get_refmet_by_name}
\alias{get_refmet_by_name}
\title{Get RefMet information by name}
\usage{
get_refmet_by_name(client, name, fields = "all")
}
\arguments{
\item{client}{mwREST client}

\item{name}{RefMet name}

\item{fields}{Fields to return (default: "all")}
}
\value{
RefMet information
}
\description{
Get RefMet information by name
}
