% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\name{get_study_data}
\alias{get_study_data}
\title{Get study data}
\usage{
get_study_data(client, study_id)
}
\arguments{
\item{client}{mwREST client}

\item{study_id}{Study ID}
}
\value{
Study data
}
\description{
Get study data
}
