% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{get_study_factors}
\alias{get_study_factors}
\title{Get study experimental factors}
\usage{
get_study_factors(client, study_id)
}
\arguments{
\item{client}{An mw_rest_client object}

\item{study_id}{Character. Study ID}
}
\value{
Tibble containing experimental factors
}
\description{
Retrieve the experimental factors (sample metadata) for a specific study.
}
\examples{
\dontrun{
client <- mw_rest_client()
factors <- get_study_factors(client, "ST000001")
}

}
