% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{get_study_metabolites}
\alias{get_study_metabolites}
\title{Get study metabolite list}
\usage{
get_study_metabolites(client, study_id)
}
\arguments{
\item{client}{An mw_rest_client object}

\item{study_id}{Character. Study ID}
}
\value{
Tibble containing metabolite information
}
\description{
Retrieve the list of metabolites measured in a specific study.
}
\examples{
\dontrun{
client <- mw_rest_client()
metabolites <- get_study_metabolites(client, "ST000001")
}

}
