% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{get_study_summary}
\alias{get_study_summary}
\title{Get study summary}
\usage{
get_study_summary(client, study_id = "ST", format = "json")
}
\arguments{
\item{client}{mwREST client}

\item{study_id}{Study ID (use "ST" for all studies)}

\item{format}{Output format ("json" or "txt")}
}
\value{
Study summary
}
\description{
Get study summary
}
