% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_endpoints}
\alias{list_endpoints}
\title{List available API endpoints}
\usage{
list_endpoints(client)
}
\arguments{
\item{client}{An mw_rest_client object (used for method dispatch)}
}
\value{
Invisibly returns NULL (called for side effects)
}
\description{
Display information about available endpoints in the mwbenchr package
organized by functional category.
}
\examples{
\dontrun{
client <- mw_rest_client()
list_endpoints(client)
}

}
