% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwbenchr.R
\docType{package}
\name{mwbenchr-package}
\alias{mwbenchr-package}
\alias{mwbenchr}
\title{mwbenchr: R Interface to Metabolomics Workbench REST API}
\description{
This package provides functions to interact with the Metabolomics
Workbench REST API (v1.2). It supports contexts: study, compound,
RefMet, MetStat, gene, protein, and moverz.
}
\section{Main Functions}{

\describe{
\item{\code{\link{mw_rest_client}}}{Initialize REST client}
\item{\code{\link{get_study_summary}}}{Get study information}
\item{\code{\link{get_compound_by_regno}}}{Get compound data}
\item{\code{\link{search_metstat}}}{Search studies by criteria}
\item{\code{\link{get_refmet_by_name}}}{Get RefMet standardized names}
}
}

\seealso{
\url{https://www.metabolomicsworkbench.org/tools/mw_rest.php}
}
\author{
\strong{Maintainer}: Dany Mukesha \email{danymukesha@gmail.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})

}
\keyword{package}
