% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwbenchr.R
\docType{package}
\name{mwbenchr}
\alias{mwbenchr-package}
\alias{mwbenchr}
\title{mwbenchr: R Interface for Direct Access to Metabolomics Workbench REST API}
\description{
The \strong{mwbenchr} package enables programmatic retrieval and processing
of metabolomics data through the Metabolomics Workbench REST API (v1.2).
It provides access to study metadata, compound information,
RefMet standardized metabolite names, metabolomics statistics (MetStat),
gene and protein data, and pathway mappings (moverz), facilitating
reproducible workflows and data integration for metabolomics research.
}
\section{Main Functions}{

\describe{
\item{\code{\link{mw_rest_client}}}{Initialize REST client}
\item{\code{\link{get_study_summary}}}{Get study information}
\item{\code{\link{get_compound_by_regno}}}{Get compound data}
\item{\code{\link{search_metstat}}}{Search studies by criteria}
\item{\code{\link{get_refmet_by_name}}}{Get RefMet standardized names}
}
}

\seealso{
\url{https://www.metabolomicsworkbench.org/tools/mw_rest.php}
}
\author{
\strong{Maintainer}: Dany Mukesha \email{danymukesha@gmail.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})

}
\keyword{package}
