% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_mw_output}
\alias{parse_mw_output}
\title{Parse a tab-delimited string into a data frame}
\usage{
parse_mw_output(result)
}
\arguments{
\item{result}{A character string, typically the output of a \code{mw_request()}
function that returns a tab-delimited string.}
}
\value{
A data frame where each row represents a set of chemical data parsed
from the input string. The columns include "Input m/z", "Matched m/z",
"Delta", "Name", "Systematic name", "Formula", "Ion", "Category", "Main
class", and "Sub class".
}
\description{
This function parses a tab-delimited string containing chemical data and
converts it into a data frame. The string contains columns such as "Input
m/z", "Matched m/z", "Delta", "Name", "Systematic name", "Formula", "Ion",
"Category", "Main class", and "Sub class".
}
\examples{
result <- paste0(
    "Input m/z\tMatched m/z\tDelta\tName\tSystematic name\tFormula\t",
    "Ion\tCategory\tMain class\tSub class\n",
    "180.063\t180.0634\t.0004\t1,3,4,5,6-Pentahydroxyhexan-2-One\t",
    "1,3,4,5,6-pentahydroxyhexan-2-one\tC6H12O6\tNeutral\tOrganic oxygen ",
    "compounds\tOrganooxygen compounds\tCarbohydrates and carbohydrate ",
    "conjugates"
)
parsed_data <- parse_mw_output(result)
print(parsed_data)

}
