% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{response_to_df}
\alias{response_to_df}
\title{Convert API response to data frame}
\usage{
response_to_df(response)
}
\arguments{
\item{response}{API response}
}
\value{
Data frame
}
\description{
Convert API response to data frame
}
