% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moverz.R
\name{search_by_mass}
\alias{search_by_mass}
\title{Search compounds by mass}
\usage{
search_by_mass(client, db, mz, ion_type, tolerance, format = "json")
}
\arguments{
\item{client}{mwREST client}

\item{db}{Database ("MB", "LIPIDS", or "REFMET")}

\item{mz}{Mass-to-charge ratio}

\item{ion_type}{Ion type (e.g., "M+H", "M-H")}

\item{tolerance}{Mass tolerance in Daltons}

\item{format}{Output format ("json" or "txt")}
}
\value{
Matching compounds
}
\description{
Search compounds by mass
}
