% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refmet.R
\name{standardize_to_refmet}
\alias{standardize_to_refmet}
\title{Standardize metabolite name to RefMet}
\usage{
standardize_to_refmet(client, name)
}
\arguments{
\item{client}{mwREST client}

\item{name}{Metabolite name to standardize}
}
\value{
Standardized RefMet name
}
\description{
Standardize metabolite name to RefMet
}
