% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moverz.R
\name{calculate_exact_mass}
\alias{calculate_exact_mass}
\title{Calculate exact mass}
\usage{
calculate_exact_mass(client, lipid_abbrev, ion_type)
}
\arguments{
\item{client}{mwREST client}

\item{lipid_abbrev}{Lipid abbreviation (e.g., "PC(34:1)")}

\item{ion_type}{Ion type (e.g., "M+H")}
}
\value{
Exact mass
}
\description{
Calculate exact mass
}
