% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten_entry}
\alias{flatten_entry}
\title{Flatten a single metabolite entry (with nested DATA field)}
\usage{
flatten_entry(entry)
}
\arguments{
\item{entry}{A list of metabolite metadata and nested sample data}
}
\value{
A single-row tibble
}
\description{
Flatten a single metabolite entry (with nested DATA field)
}
