% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refmet.R
\name{get_all_refmet_names}
\alias{get_all_refmet_names}
\title{Get all RefMet names}
\usage{
get_all_refmet_names(client)
}
\arguments{
\item{client}{An mw_rest_client object}
}
\value{
Tibble containing all RefMet names
}
\description{
Retrieve the complete list of standardized metabolite names from RefMet.
This can be large, so consider using caching.
}
\examples{
\dontrun{
# Use caching for this large dataset
client <- mw_rest_client(cache = TRUE)
all_names <- get_all_refmet_names(client)
}

}
