% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound.R
\name{get_compound_by_pubchem_cid}
\alias{get_compound_by_pubchem_cid}
\title{Get compound information by PubChem CID}
\usage{
get_compound_by_pubchem_cid(client, cid, fields = "all", format = "json")
}
\arguments{
\item{client}{mwREST client}

\item{cid}{PubChem CID}

\item{fields}{Fields to return (default: "all")}

\item{format}{Output format ("json" or "txt")}
}
\value{
Compound information
}
\description{
Get compound information by PubChem CID
}
