% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound.R
\name{get_compound_by_regno}
\alias{get_compound_by_regno}
\title{Get compound information by registry number}
\usage{
get_compound_by_regno(client, regno, fields = "all", format = "json")
}
\arguments{
\item{client}{mwREST client}

\item{regno}{Registry number}

\item{fields}{Fields to return (default: "all")}

\item{format}{Output format ("json" or "txt")}
}
\value{
Compound information
}
\description{
Get compound information by registry number
}
