% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refmet.R
\name{get_refmet_by_name}
\alias{get_refmet_by_name}
\title{Get RefMet information by name}
\usage{
get_refmet_by_name(client, name, fields = "all")
}
\arguments{
\item{client}{An mw_rest_client object}

\item{name}{Character. RefMet metabolite name}

\item{fields}{Character. Fields to return (default: "all")}
}
\value{
Tibble containing RefMet information
}
\description{
Retrieve RefMet (Reference list of Metabolite names) information
for a specific metabolite name.
}
\examples{
\dontrun{
client <- mw_rest_client()
refmet_info <- get_refmet_by_name(client, "Cholesterol")
}

}
