% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\name{get_study_factors}
\alias{get_study_factors}
\title{Get study factors}
\usage{
get_study_factors(client, study_id)
}
\arguments{
\item{client}{mwREST client}

\item{study_id}{Study ID}
}
\value{
Study factors
}
\description{
Get study factors
}
