% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{get_study_summary}
\alias{get_study_summary}
\title{Get study summary information}
\usage{
get_study_summary(client, study_id = "ST", format = "json")
}
\arguments{
\item{client}{An mw_rest_client object}

\item{study_id}{Character. Study ID (use "ST" for all studies)}

\item{format}{Character. Output format ("json" or "txt")}
}
\value{
Study summary data as tibble (JSON) or character string (txt)
}
\description{
Retrieve summary information for one or all studies in the Metabolomics
Workbench database.
}
\examples{
\dontrun{
client <- mw_rest_client()

# Get all studies
all_studies <- get_study_summary(client)

# Get specific study
study_info <- get_study_summary(client, study_id = "ST000001")
}

}
