% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_endpoints}
\alias{list_endpoints}
\title{Print available endpoints}
\usage{
list_endpoints(client)
}
\arguments{
\item{client}{mwREST client}
}
\value{
Information about available endpoints
}
\description{
Print available endpoints
}
