% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{mw_request}
\alias{mw_request}
\title{Make a request to the Metabolomics Workbench REST API}
\usage{
mw_request(client, endpoint, parse = TRUE, format = "json", ...)
}
\arguments{
\item{client}{mwREST client object}

\item{endpoint}{API endpoint path (e.g., "compound/regno/11/name")}

\item{parse}{Should the response be parsed? (TRUE/FALSE)}

\item{format}{Output format ("json" or "txt")}

\item{...}{Additional parameters passed to httr2::req_perform}
}
\value{
API response
}
\description{
Make a request to the Metabolomics Workbench REST API
}
\keyword{internal}
