% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwbenchr.R
\docType{package}
\name{mwbenchr}
\alias{mwbenchr-package}
\alias{mwbenchr}
\title{mwbenchr: R Interface to Metabolomics Workbench REST API}
\description{
This package provides functions to interact with the Metabolomics
Workbench REST API (v1.2). It supports contexts: study, compound,
refmet, metstat, gene, protein, and moverz.
}
\seealso{
\itemize{
 \item Visit \url{https://www.metabolomicsworkbench.org/tools/mw_rest.php}
 for details of the API.
}
}
\author{
\strong{Maintainer}: Dany Mukesha \email{danymukesha@gmail.com} (\href{https://orcid.org/0009-0001-9514-751X}{ORCID})

}
\keyword{internal}
