% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{print.mw_rest_client}
\alias{print.mw_rest_client}
\title{Print method for mw_rest_client}
\usage{
\method{print}{mw_rest_client}(x, ...)
}
\arguments{
\item{x}{An mw_rest_client object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for mw_rest_client
}
