% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{response_to_df}
\alias{response_to_df}
\title{Convert mwbenchr API response into a clean data frame}
\usage{
response_to_df(response)
}
\arguments{
\item{response}{A parsed response from mwbenchr}
}
\value{
A tibble
}
\description{
Handles responses like:
}
\details{
- Flat named list (e.g. compound, study)
- Named list of rows (e.g. search results from search_metstat)
- List of metabolite entries (with nested DATA)
}
